Stress/Stress/requirements.txt
tensorflow>=2.12.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.0.0
matplotlib>=3.7.0
spotipy>=2.23.0
python-dotenv>=1.0.0
